chore: move files to src
