+++ index.js
@@ -6,7 +6,7 @@ const act_on_approve_command = require('./src/act_on_approve_command');
 const act_on_update_command = require('./src/act_on_update_command');
 const act_on_assign_command = require("./src/act_on_unassign_command");
 const act_on_unassign_command = require("./src/act_on_unassign_command");
-const auto_add_default_labels_and_lock_new_issues = require('./auto_add_default_labels_and_lock_new_issues');
+const auto_add_default_labels_and_lock_new_issues = require('./src/auto_add_default_labels_and_lock_new_issues');
 
 const run = async () => {
     const token = core.getInput('token', { required: true });
