+++ src/auto_add_default_labels_and_lock_new_issues.js
